Working on getting through object-oriented coding
Having a really tough time seeing how to apply this practically, as well as arrays, because it is an abstract way of thinking that I'm not familiar with. It will be okay, I just need to do some more exercises that use these techniques and become familiar with the structure of both.