Created team class
Will figure out how it follows the ball and give it a catching method (use distance). Once this is done, I will add it to the main script. Afterwards, I will create a rival class (basically copy paste). I will then make them unable to collide, and add a state so that when the ball is caught, the teammate having caught it will be controlled by the user using the mouse.